Research/encoding periodic (#12502)
* rebase v1.0.0

* add pure period

* type change

* more replace

* achieve several functions

* add fft method

* add new function

* complete encode

* decode

* doc

+++ library-udf/src/main/java/org/apache/iotdb/library/dprofile/UDTFSpline.java
@@ -90,7 +90,7 @@ public class UDTFSpline implements UDTF {
                             / (samplePoints - 1)
                         + 0.5);
         double yhead = psf.value(approximation);
-        collector.putDouble(minimumTimestamp + approximation, yhead);
+        collector.putDouble(minimumTimestamp + (long) approximation, yhead);
       }
     }
   }
